fix printf call with sizeof(time_t) != sizeof(long)
In environments with 32-bit long and 64-bit time_t (e.g. the x32 ABI
for x86) there is a type mismatch between the format string and its
argument. For now, always cast the timestamp to long. (Postponing this
problem until 2038 where 32-bit time_t will overflow anyway.)

+++ perf_examples/self_smpl_multi.c
@@ -165,7 +165,7 @@ do_cycles(void)
 		gettimeofday(&now, NULL);
 		if (now.tv_sec > last.tv_sec) {
 			printf("%ld: myid = %3d, fd = %3d, count = %4ld, iter = %4ld, rate = %ld/Kiter\n",
-				now.tv_sec - start.tv_sec,
+				(long)(now.tv_sec - start.tv_sec),
 				myid,
 				fd2ov[myid].fd,
 				count[myid], iter[myid],
